Landscape page work
Fixed a bug that was showing the corner of a border under the header
buttons on all pages.

Added a picture to the planting section.

Designed the mulching section and added a mulch calculator.

Changed some links on the homepage from relative to absolute.

Changed a seasonal message on the homepage in the lawn mowing section.